requirements.txt
python-vlc==3.0.18121
Pillow==10.2.0
